Add files via upload
This is a simple shell script that when runned on a unix terminal (use the macOS X terminal for better results), using sh command, will list all the processes running on the computer on the moment the command was executed, in a file on the pprints folder (the files are numbered). For the program work properly you need to place all the files on the pprint folder on the home directory.